Cybersecurity Concerns in Enterprise IoT.
IoT is a new technology opportunity that has developed in the last few years. Today, there are currently 10 billion interconnected devices and this is expected to grow to 50 billion by 2020, and that is only 3 years away. With that many devices it makes for a highly vulnerable network with billions of vulnerabilities that can be exploited by determined hackers. Today, we are using smartphones as a swiss army knife, and we are embracing smart homes, smart cars, self-driving cars, smart factories, smart hospitals, smart airports, smart cities, etc and the list is growing bigger day by day. In evaluating security needs for enterprise IoT, I will have look at the entire technology stack which is comprised of three different areas: IT, OT and Devices.